+++ btclib/bms.py
@@ -141,8 +141,7 @@ from .alias import BMSig, BMSigTuple, PrvKey, String
 from .base58address import h160_from_b58address, p2pkh, p2wpkh_p2sh
 from .base58wif import wif_from_prvkey
 from .bech32address import p2wpkh, witness_from_b32address
-from .curvemult import mult
-from .curves import secp256k1
+from .curve import mult, secp256k1
 from .network import NETWORKS
 from .secpoint import bytes_from_point
 from .to_prvkey import prvkeyinfo_from_prvkey
