+++ btclib/dh.py
@@ -23,9 +23,7 @@ from hashlib import sha256
 from typing import Any, Callable
 
 from .alias import HashF, Point
-from .curve import Curve
-from .curvemult import mult
-from .curves import secp256k1
+from .curve import Curve, mult, secp256k1
 
 KDF = Callable[[bytes, int, Curve, HashF], Any]
