+++ btclib/rfc6979.py
@@ -39,8 +39,7 @@ import hmac
 from hashlib import sha256
 
 from .alias import HashF, Octets, PrvKey, String
-from .curve import Curve
-from .curves import secp256k1
+from .curve import Curve, secp256k1
 from .hashes import reduce_to_hlen
 from .to_prvkey import int_from_prvkey
 from .utils import bytes_from_octets, int_from_bits
